+++ src/Herramientas/Desgastador/Desgatador.java
@@ -2,5 +2,5 @@ package Herramientas.Desgastador;
 
 public interface Desgatador {
 
-    int calcularDesgaste();
+    float calcularDesgaste();
 }
