+++ test/DesgastadoPorUsosTest.java
@@ -8,14 +8,14 @@ public class DesgastadoPorUsosTest {
     @Test
     void desgastadorPorUsosSeCreaCorractamente(){
         DesgatadorPorUsos desgatadorPorUsos = new DesgatadorPorUsos(2,5);
-        int usosRestante = desgatadorPorUsos.usosRestante();
+        float usosRestante = desgatadorPorUsos.usosRestante();
         assertEquals(2,usosRestante);
     }
 
     @Test
     void desgastadorPorUsosSeUsaTotosSusUsosSuDurabilidadEsCero(){
         DesgatadorPorUsos desgatadorPorUsos = new DesgatadorPorUsos(3,5);
-        int desgastado = 0;
+        float desgastado = 0;
         for (int i=0; i < 4; i++){
             desgastado = desgatadorPorUsos.calcularDesgaste();
         }
