+++ test/PicoTest.java
@@ -8,8 +8,8 @@ public class PicoTest {
     @Test
     void seCreaUnPicoDeMaderaCorrectamente(){
         PicoMadera picoMadera = new PicoMadera();
-        int durabilidad = picoMadera.getDurabilidad();
-        int fuerza = picoMadera.getFuerza();
+        float durabilidad = picoMadera.getDurabilidad();
+        float fuerza = picoMadera.getFuerza();
         assertEquals(100,durabilidad);
         assertEquals(2,fuerza);
     }
@@ -17,7 +17,7 @@ public class PicoTest {
     @Test
     void seCreaUnPicoMaderaYDesgastaCorrectamente(){
         PicoMadera picoMadera = new PicoMadera();
-        int durabilidad = picoMadera.getDurabilidad();
+        float durabilidad = picoMadera.getDurabilidad();
         PicoMadera.desgastar();
         assertEquals(durabilidad-2, picoMadera.getDurabilidad());
     }
@@ -25,8 +25,8 @@ public class PicoTest {
     @Test
     void seCreaUnPicoDePiedraCorrectamente(){
         PicoPiedra picoPiedra = new PicoPiedra();
-        int durabilidad = picoPiedra.getDurabilidad();
-        int fuerza = picoPiedra.getFuerza();
+        float durabilidad = picoPiedra.getDurabilidad();
+        float fuerza = picoPiedra.getFuerza();
         assertEquals(200,durabilidad);
         assertEquals(4,fuerza);
     }
@@ -34,7 +34,7 @@ public class PicoTest {
     @Test
     void seCreaUnPicoDePiedraYDesgastaCorrectamente(){
         PicoPiedra picoPiedra = new PicoPiedra();
-        int durabilidad = picoPiedra.getDurabilidad();
+        float durabilidad = picoPiedra.getDurabilidad();
         PicoMadera.desgastar();
         assertEquals(durabilidad-2, picoPiedra.getDurabilidad());
     }
@@ -42,8 +42,8 @@ public class PicoTest {
     @Test
     void seCreaUnPicoMetalYDesgastaCorrectamente(){
         PicoMetal picoMetal = new PicoMetal();
-        int durabilidad = picoMetal.getDurabilidad();
-        int fuerza = picoMetal.getFuerza();
+        float durabilidad = picoMetal.getDurabilidad();
+        float fuerza = picoMetal.getFuerza();
         assertEquals(400,durabilidad);
         assertEquals(12,fuerza);
     }
@@ -51,7 +51,7 @@ public class PicoTest {
     @Test
     void seCreaUnPicoDePiedraSeUsaYNoCambiaDurabilidad(){
         PicoMetal picoMetal = new PicoMetal();
-        int durabilidad = picoMetal.getDurabilidad();
+        float durabilidad = picoMetal.getDurabilidad();
         PicoMadera.desgastar();
         assertEquals(durabilidad, picoMetal.getDurabilidad());
     }
@@ -59,8 +59,8 @@ public class PicoTest {
     @Test
     void seCreaUnPicoFinoCorrectamnete(){
         PicoFino picoFino = new PicoFino();
-        int durabilidad = picoFino.getDurabilidad();
-        int fuerza = picoFino.getFuerza();
+        float durabilidad = picoFino.getDurabilidad();
+        float fuerza = picoFino.getFuerza();
         assertEquals(1000,durabilidad);
         assertEquals(20,fuerza);
     }
@@ -68,7 +68,7 @@ public class PicoTest {
     @Test
     void seCreaPicoFinoYSedesgastaCorrectamente(){
         PicoFino picoFino = new PicoFino();
-        int durabilidad = picoFino.getDurabilidad();
+        float durabilidad = picoFino.getDurabilidad();
         picoFino.desgastar();
         assertEquals(durabilidad - (durabilidad/10), picoFino.getDurabilidad());
     }
